requirements.txt
discord.py
python-dotenv
gTTS
requests